added modules server and core
