+++ client/src/main/java/Client.java
@@ -1,0 +1,2 @@
+public class Client {
+}
