+++ core/src/main/java/io/deeplay/domain/Color.java
