+++ core/src/main/java/io/deeplay/domain/GameType.java
@@ -1,0 +1,7 @@
+package io.deeplay.domain;
+
+public enum GameType {
+    BotVsBot,
+    HumanVsBot,
+    HumanVsHuman
+}
