+++ core/src/main/java/io/deeplay/model/piece/Bishop.java
