+++ core/src/main/java/io/deeplay/model/piece/Empty.java
