+++ core/src/main/java/io/deeplay/model/piece/King.java
