+++ core/src/main/java/io/deeplay/model/piece/Queen.java
