+++ core/src/main/java/io/deeplay/model/player/Bot.java
