+++ core/src/test/java/io/deeplay/model/CoordinatesTest.java
