+++ core/src/test/java/io/deeplay/model/piece/BishopTest.java
