+++ core/src/test/java/io/deeplay/model/piece/EmptyTest.java
