+++ core/src/test/java/io/deeplay/model/piece/KingTest.java
