+++ core/src/test/java/io/deeplay/model/piece/KnightTest.java
