+++ core/src/test/java/io/deeplay/model/piece/QueenTest.java
