+++ settings.gradle.kts
@@ -1,1 +1,5 @@
 rootProject.name = "chess-wars"
+
+include("core")
+include("server")
+include("client")
