--- src/main/java/io/deeplay/domain/GameType.java
@@ -1,7 +1,0 @@
-package io.deeplay.domain;
-
-public enum GameType {
-    BotVsBot,
-    HumanVsBot,
-    HumanVsHuman
-}
