--- src/main/java/io/deeplay/engine/GameState.java
@@ -1,10 +1,0 @@
-package io.deeplay.engine;
-
-import io.deeplay.model.Board;
-
-public class GameState {
-    public static boolean check(Board board) {
-
-        return false;
-    }
-}
